deduplicated a lot of code
