+++ src/rust/wpcomics/sources/comiconlinefree/src/helper.rs
@@ -1,7 +1,6 @@
 use aidoku::{
 	prelude::format,
 	std::{String, StringRef, Vec},
-	MangaStatus,
 };
 use wpcomics_template::helper::urlencode;
 
@@ -13,32 +12,11 @@ pub fn listing_mapping(listing: String) -> String {
 	})
 }
 
-pub fn status_map(arg1: String) -> MangaStatus {
-	return match arg1.as_str() {
-		"Ongoing" => MangaStatus::Ongoing,
-		"Completed" => MangaStatus::Completed,
-		_ => MangaStatus::Unknown,
-	};
-}
-
 pub fn convert_time(time_ago: String) -> f64 {
 	let time_object = StringRef::from(time_ago).0;
 	return time_object
 		.as_date("MM/dd/yy", Some("en_US"), None)
 		.unwrap_or(0.0);
-}
-
-pub fn trunc_trailing_comic(title: String) -> String {
-	let temp = title.chars().rev().collect::<String>();
-	if temp.find("cimoC") == Some(0) {
-		return temp
-			.replacen("cimoC", "", 1)
-			.chars()
-			.rev()
-			.collect::<String>();
-	} else {
-		return temp.chars().rev().collect::<String>();
-	}
 }
 
 pub fn get_search_url(
