+++ src/rust/wpcomics/sources/nettruyen/src/helper.rs
@@ -1,13 +1,13 @@
 use aidoku::{
+	prelude::format,
 	std::Vec,
 	std::{current_date, String, StringRef},
 	MangaStatus,
 };
-use wpcomics_template::helper::i32_to_string;
 
 // MARK: Mappings
 pub fn get_tag_id(genre: String) -> String {
-	let id = match genre.as_str() {
+	String::from(match genre.as_str() {
 		"Action" => "1",
 		"Adult" => "2",
 		"Adventure" => "3",
@@ -66,8 +66,7 @@ pub fn get_tag_id(genre: String) -> String {
 		"16+" => "60",
 		"18+" => "61",
 		_ => "",
-	};
-	return String::from(id);
+	})
 }
 
 pub fn status_map(arg1: String) -> MangaStatus {
@@ -90,43 +89,22 @@ pub fn get_search_url(
 	completed: i32,
 	chapter_count: i32,
 ) -> String {
-	let mut url = String::new();
-	url.push_str(&base_url);
-	if query.len() > 0 {
-		url.push_str("/tim-truyen?page=");
-		url.push_str(i32_to_string(page).as_str());
-		url.push_str("&keyword=");
-		url.push_str(&query);
-		return url;
-	}
-	url.push_str("/tim-truyen-nang-cao?");
-	url.push_str("genres=");
-	if include.len() > 0 {
-		url.push_str(include.join(",").as_str());
-	}
-	url.push_str("&notgenres=");
-	if exclude.len() > 0 {
-		url.push_str(exclude.join(",").as_str());
-	}
-	url.push_str("&gender=");
-	url.push_str(i32_to_string(gender).as_str());
-	url.push_str("&status=");
-	url.push_str(i32_to_string(completed).as_str());
-	url.push_str("&minchapter=");
-	url.push_str(i32_to_string(chapter_count).as_str());
-	url.push_str("&sort=");
-	url.push_str(i32_to_string(sort_by).as_str());
-	url.push_str("&page=");
-	url.push_str(i32_to_string(page).as_str());
-	return url;
+	return if query.len() > 0 {
+		format!("{base_url}/tim-truyen?page={page}&keyword={query}")
+	} else {
+		format!(
+			"{base_url}/tim-truyen-nang-cao?genres={}&notgenres={}&gender={gender}&status={completed}&minchapter={chapter_count}&sort={sort_by}&page={page}",
+			include.join(","),
+			exclude.join(",")
+		)
+	};
 }
 
 pub fn convert_time(time_ago: String) -> f64 {
-	let time: f64;
 	let current_time = current_date();
 	let time_arr = time_ago.split(" ").collect::<Vec<&str>>();
 	if time_arr.len() > 2 {
-		time = match time_arr[1] {
+		return match time_arr[1] {
 			"giây" => current_time - time_arr[0].parse::<f64>().unwrap_or(0.0),
 			"phút" => current_time - time_arr[0].parse::<f64>().unwrap_or(0.0) * 60.0,
 			"giờ" => current_time - time_arr[0].parse::<f64>().unwrap_or(0.0) * 3600.0,
@@ -134,21 +112,20 @@ pub fn convert_time(time_ago: String) -> f64 {
 			_ => current_time,
 		};
 	} else if String::from(time_arr[0]) == time_ago {
-		let time_object = StringRef::from(time_ago).0;
-		time = time_object
+		return StringRef::from(time_ago)
+			.0
 			.as_date("dd/MM/yy", Some("en_US"), Some("Asia/Ho_Chi_Minh"))
 			.unwrap_or(0.0);
 	} else {
-		let current_year: i32 = 1970 + (current_time / 31536000.0) as i32;
-		let mut modified_time = String::from(time_arr[0]);
-		modified_time.push_str(" ");
-		modified_time.push_str(time_arr[1]);
-		modified_time.push_str("/");
-		modified_time.push_str(&i32_to_string(current_year));
-		let time_object = StringRef::from(modified_time).0;
-		time = time_object
+		let modified_time = format!(
+			"{} {}/{}",
+			time_arr[0],
+			time_arr[1],
+			1970 + (current_time / 31536000.0) as i32
+		);
+		return StringRef::from(modified_time)
+			.0
 			.as_date("HH:mm dd/MM/yyyy", Some("en_US"), Some("Asia/Ho_Chi_Minh"))
 			.unwrap_or(0.0);
 	}
-	return time;
 }
