--- src/rust/wpcomics/sources/xoxocomics/res/filters.json
@@ -1,66 +1,0 @@
-[
-	{
-		"type": "title"
-	},
-	{
-		"type": "select",
-		"name": "Genre",
-		"options": [
-			"None",
-			"Marvel",
-			"DC Comics",
-			"Action",
-			"Adventure",
-			"Anthology",
-			"Anthropomorphic",
-			"Biography",
-			"Children",
-			"Comedy",
-			"Crime",
-			"Cyborgs",
-			"Dark Horse",
-			"Demons",
-			"Drama",
-			"Fantasy",
-			"Family",
-			"Fighting",
-			"Gore",
-			"Graphic Novels",
-			"Historical",
-			"Horror",
-			"Leading Ladies",
-			"Literature",
-			"Magic",
-			"Manga",
-			"Martial Arts",
-			"Mature",
-			"Mecha",
-			"Military",
-			"Movie Cinematic Link",
-			"Mystery",
-			"Mythology",
-			"Psychological",
-			"Personal",
-			"Political",
-			"Post-Apocalyptic",
-			"Pulp",
-			"Robots",
-			"Romance",
-			"Sci-Fi",
-			"Slice of Life",
-			"Science Fiction",
-			"Sport",
-			"Spy",
-			"Superhero",
-			"Supernatural ",
-			"Suspense",
-			"Thriller",
-			"Vampires",
-			"Vertigo",
-			"Video Games",
-			"War",
-			"Western",
-			"Zombies"
-		]
-	}
-]
