+++ src/rust/wpcomics/template/src/helper.rs
@@ -1,11 +1,16 @@
-use aidoku::{prelude::format, std::ArrayRef, std::String, std::Vec, MangaStatus};
+use aidoku::{prelude::format, std::String, std::Vec};
 
-#[macro_export]
-macro_rules! scan {
-	( $string:expr, $sep:expr, $( $x:ty ),+ ) => {{
-		let mut iter = $string.split($sep);
-		($(iter.next().and_then(|word| word.parse::<$x>().ok()),)*)
-	}}
+pub fn trunc_trailing_comic(title: String) -> String {
+	let temp = title.chars().rev().collect::<String>();
+	if temp.find("cimoC") == Some(0) {
+		return temp
+			.replacen("cimoC", "", 1)
+			.chars()
+			.rev()
+			.collect::<String>();
+	} else {
+		return temp.chars().rev().collect::<String>();
+	}
 }
 
 pub fn extract_f32_from_string(title: String, text: String) -> f32 {
@@ -55,102 +60,62 @@ pub fn urlencode(string: String) -> String {
 	String::from_utf8(result).unwrap_or(String::new())
 }
 
-pub fn i32_to_string(mut integer: i32) -> String {
-	if integer == 0 {
-		return String::from("0");
-	}
-	let mut string = String::with_capacity(11);
-	let pos = if integer < 0 {
-		string.insert(0, '-');
-		1
-	} else {
-		0
-	};
-	while integer != 0 {
-		let mut digit = integer % 10;
-		if pos == 1 {
-			digit *= -1;
-		}
-		string.insert(pos, char::from_u32((digit as u32) + ('0' as u32)).unwrap());
-		integer /= 10;
-	}
-	return string;
-}
-
-pub fn join_string_array(array: ArrayRef, delimeter: String) -> String {
-	let mut string = String::new();
-	let mut at = 0;
-	for item in array {
-		if at != 0 {
-			string.push_str(&delimeter);
-		}
-		string.push_str(item.as_node().text().read().as_str());
-		at += 1;
-	}
-	return string;
-}
-
-pub fn status_from_string(status: String) -> MangaStatus {
-	if status == "Ongoing" {
-		return MangaStatus::Ongoing;
-	} else if status == "Completed" {
-		return MangaStatus::Completed;
-	} else if status == "Hiatus" {
-		return MangaStatus::Hiatus;
-	} else if status == "Cancelled" {
-		return MangaStatus::Cancelled;
-	} else {
-		return MangaStatus::Unknown;
-	}
-}
-
-pub fn is_numeric_char(c: char) -> bool {
-	return (c >= '0' && c <= '9') || c == '.';
-}
-
-pub fn get_chapter_number(id: String) -> f32 {
-	let mut number_string = String::new();
-	let mut i = id.len() - 1;
-	for c in id.chars().rev() {
-		if !is_numeric_char(c) {
-			number_string = String::from(&id[i + 1..]);
-			break;
-		}
-		i -= 1;
-	}
-	if number_string.len() == 0 {
-		return 0.0;
-	}
-	return number_string.parse::<f32>().unwrap_or(0.0);
-}
-
-pub fn string_replace(string: String, search: String, replace: String) -> String {
-	let mut result = String::new();
-	let mut at = 0;
-	for c in string.chars() {
-		if c == search.chars().next().unwrap() {
-			if string[at..].starts_with(&search) {
-				result.push_str(&replace);
-				at += search.len();
-			} else {
-				result.push(c);
-			}
-		} else {
-			result.push(c);
-		}
-		at += 1;
-	}
-	return result;
-}
-
-pub fn stupidencode(string: String) -> String {
-	let mut result = String::new();
-	for c in string.chars() {
-		if c.is_alphanumeric() {
-			result.push(c);
-		} else if c == ' ' {
-			result.push('_');
-		}
-	}
-	return result;
+pub fn get_tag_id(genre: i64) -> String {
+	return String::from(match genre {
+		1 => "marvel",
+		2 => "dc-comics",
+		3 => "action",
+		4 => "adventure",
+		5 => "anthology",
+		6 => "anthropomorphic",
+		7 => "biography",
+		8 => "children",
+		9 => "comedy",
+		10 => "crime",
+		11 => "cyborgs",
+		12 => "dark-horse",
+		13 => "demons",
+		14 => "drama",
+		15 => "fantasy",
+		16 => "family",
+		17 => "fighting",
+		18 => "gore",
+		19 => "graphic-novels",
+		20 => "historical",
+		21 => "horror",
+		22 => "leading-ladies",
+		23 => "literature",
+		24 => "magic",
+		25 => "manga",
+		26 => "martial-arts",
+		27 => "mature",
+		28 => "mecha",
+		29 => "military",
+		30 => "movie-cinematic-link",
+		31 => "mystery",
+		32 => "mythology",
+		33 => "psychological",
+		34 => "personal",
+		35 => "political",
+		36 => "post-apocalyptic",
+		37 => "pulp",
+		38 => "robots",
+		39 => "romance",
+		40 => "sci-fi",
+		41 => "slice-of-life",
+		42 => "science-fiction",
+		43 => "sport",
+		44 => "spy",
+		45 => "superhero",
+		46 => "supernatural",
+		47 => "suspense",
+		48 => "thriller",
+		49 => "vampires",
+		50 => "vertigo",
+		51 => "video-games",
+		52 => "war",
+		53 => "western",
+		54 => "zombies",
+		_ => "",
+	});
 }
